# fe/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "my-fam-tree API",
    "description": "Family platform API",
    "license": {
      "name": "BUSL-1.1",
      "identifier": "BUSL-1.1"
    },
    "version": "0.1.4"
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin"
    }
  ],
  "paths": {
...
    "/api/v1/families/{family_id}/members": {
      "get": {
        "tags": [
          "members"
        ],
        "operationId": "members_list",
        "parameters": [
          {
            "name": "family_id",
            "in": "path",
            "description": "Family id (must match the active X-Family-Id)",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Members list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MembersListResponseBody"
                }
              }
            }
          },
          "401": {
            "description": "Path family_id does not match active family"
          },
          "403": {
            "description": "Insufficient role (admin / owner required)"
          }
        },
        "security": [
          {
            "cookie_access": []
          }
        ]
      }
    },
    "/api/v1/families/{family_id}/members/{user_id}": {
      "delete": {
        "tags": [
          "members"
        ],
        "operationId": "members_revoke",
        "parameters": [
          {
            "name": "family_id",
            "in": "path",
            "description": "Family id (must match the active X-Family-Id)",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "user_id",
            "in": "path",
            "description": "Target member's user id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Member revoked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NullResponseBody"
                }
              }
            }
          },
          "401": {
            "description": "Path family_id does not match active family"
          },
          "403": {
            "description": "Role-matrix violation"
          },
          "404": {
            "description": "Target is not a member"
          }
        },
        "security": [
          {
            "cookie_access": []
          }
        ]
      },
      "patch": {
        "tags": [
          "members"
        ],
        "operationId": "members_set_role",
        "parameters": [
          {
            "name": "family_id",
            "in": "path",
            "description": "Family id (must match the active X-Family-Id)",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "user_id",
            "in": "path",
            "description": "Target member's user id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetRoleReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Role updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberResponseBody"
                }
              }
            }
          },
          "401": {
            "description": "Path family_id does not match active family"
          },
          "403": {
            "description": "Role-matrix violation"
          },
          "404": {
            "description": "Target is not a member"
          }
        },
        "security": [
          {
            "cookie_access": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MemberDto": {
        "type": "object",
        "required": [
          "user_id",
          "email",
          "display_name",
          "role",
          "joined_at"
        ],
        "properties": {
          "display_name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "joined_at": {
            "type": "string",
            "format": "date-time"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "MemberResponseBody": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/MemberDto"
          },
          "meta": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ResponseMeta"
              }
            ]
          }
        }
      },
      "MembersList": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MemberDto"
            }
          }
        }
      },
      "MembersListResponseBody": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/MembersList"
          },
          "meta": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ResponseMeta"
              }
            ]
          }
        }
      },
...
      "NullResponseBody": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {},
          "meta": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ResponseMeta"
              }
            ]
          }
        }
      },
      "Pagination": {
        "type": "object",
        "description": "Cursor-based pagination block. `next_cursor` is `None` on the last page.",
        "required": [
          "limit",
          "returned"
        ],
        "properties": {
          "limit": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "next_cursor": {
            "type": [
              "string",
              "null"
            ]
          },
          "returned": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "ResponseMeta": {
        "type": "object",
        "description": "Per-response metadata. Currently carries pagination cursors, the\n`request_id` for tracing-friendly client diagnostics, and any soft\nvalidation `warnings` attached by the handler.",
        "properties": {
          "pagination": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Pagination"
              }
            ]
          },
          "request_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "warnings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Warning"
            },
            "description": "Soft warnings. Serialized only when non-empty so the wire shape\nstays unchanged for handlers that don't produce any."
          }
        }
      },
      "Role": {
        "type": "string",
        "enum": [
          "user",
          "admin",
          "owner"
        ]
      },
      "SetRoleReq": {
        "type": "object",
        "required": [
          "role"
        ],
        "properties": {
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        },
        "additionalProperties": false
      },
...
      "Warning": {
        "type": "object",
        "description": "Soft-validation warning.\n\nSurfaced under `meta.warnings` on the success envelope when the\ncaller's input violates a heuristic rule that we don't want to reject\noutright (e.g. parent-child age gap < 14 years). The `code` is a\nstable i18n key (`warning.<snake_case>`); `path` is an optional JSON\nPointer to the relevant field.",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "path": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      }
    },
    "securitySchemes": {
      "cookie_access": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access"
      }
    }
  },
...
}